Revert "Task management + UI improvement + General Refactoring + Better logging + container improvements"
