--- .gitignore
@@ -1,18 +1,0 @@
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-
-# Environments
-.venv/
-venv/
-ENV/
-
-# VS Code
-.vscode/
-
-# OS / misc
-.DS_Store
-
-# Logs
-*.log
