+++ README.md
@@ -45,10 +45,7 @@ All wrapped in a dark, minimal interface designed to live inside Docker/Unraid.
     * Run automatically by cron
 * Logging
 
-  * Each run creates a timestamped log file at `/tasks/<slug>/logs/run_YYYYMMDD_HHMMSS.log`
-  * Main `logs.txt` accumulates all runs after completion
-  * ANSI escape sequences stripped for clean display in UI (only color formatting retained)
-  * Real-time log viewer with auto-scroll, level-based coloring, and manual pause support
+  * Each run appends stdout/stderr to the task’s `logs.txt`
   * Includes command line + exit code info
 * Media wall dashboard
 
@@ -84,7 +81,7 @@ All wrapped in a dark, minimal interface designed to live inside Docker/Unraid.
   * Status (idle / running / paused)
   * Cron expression
   * Last run time
-  * Actions (Run, Cancel, Pause/Unpause, Edit, Delete)
+  * Actions (Run, Pause/Unpause, Delete)
 * Task editor:
 
   * Task name
@@ -118,20 +115,8 @@ docker run -d \
   -e DOWNLOADS_DIR=/downloads \
   -e PUID=99 \
   -e PGID=100 \
-  # Optional login (remove these if you don't need auth)
-  -e ARTILLERY_LOGIN_REQUIRED=1 \
-  -e ARTILLERY_USERNAME=admin \
-  -e ARTILLERY_PASSWORD=change-me \
   -v /mnt/user/appdata/artillery/config:/config \
   -v /mnt/user/appdata/artillery/tasks:/tasks \
   -v /mnt/user/pictures:/downloads \
   obviousviking/artillery
 ```
-
-### Optional login (username/password)
-
-Authentication is disabled by default. To require a login page:
-
-* `ARTILLERY_LOGIN_REQUIRED=1` – enable the login screen. Omit or set to `0` to leave the app open.
-* `ARTILLERY_USERNAME` / `ARTILLERY_PASSWORD` – optional overrides for the credentials. If not provided, they default to `admin` / `artillery`.
-* You can leave username/password unset in container creation; they are only read when login is enabled.
