+++ scheduler.py
@@ -3,7 +3,12 @@ from datetime import datetime
 
 from croniter import croniter
 
-from task_runtime import ensure_data_dirs, TASKS_ROOT, read_text, run_task_background
+from app import (
+    ensure_data_dirs,
+    TASKS_ROOT,
+    read_text,
+    run_task_background,
+)
 
 
 def should_run_now(cron_expr: str, now: datetime) -> bool:
@@ -14,10 +19,7 @@ def should_run_now(cron_expr: str, now: datetime) -> bool:
     cron_expr = cron_expr.strip()
     if not cron_expr:
         return False
-    try:
-        return croniter.match(cron_expr, now)
-    except Exception:
-        return False
+    return croniter.match(cron_expr, now)
 
 
 def main():
@@ -27,16 +29,10 @@ def main():
     if not os.path.isdir(TASKS_ROOT):
         return
 
-    try:
-        entries = list(os.scandir(TASKS_ROOT))
-    except Exception:
-        entries = []
-
-    for entry in sorted(entries, key=lambda e: e.name):
-        if not entry.is_dir():
+    for slug in sorted(os.listdir(TASKS_ROOT)):
+        task_folder = os.path.join(TASKS_ROOT, slug)
+        if not os.path.isdir(task_folder):
             continue
-        slug = entry.name
-        task_folder = entry.path
 
         cron_path = os.path.join(task_folder, "cron.txt")
         cron_expr = read_text(cron_path)
@@ -56,12 +52,7 @@ def main():
 
         print(f"[scheduler] {now.isoformat()} - running task '{slug}' with cron '{cron_expr}'")
 
-        # Create lock atomically to avoid races
-        try:
-            with open(lock_path, "x"):
-                pass
-        except FileExistsError:
-            continue
+        open(lock_path, "w").close()
         run_task_background(task_folder)
