+++ templates/base.html
@@ -17,8 +17,7 @@
     <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}" />
 </head>
 
-<body class="{{ body_class | default('') }}">
-    {% if (not login_required) or session.get('authenticated') %}
+<body>
     <nav class="navbar navbar-expand-lg navbar-dark navbar-artillery mb-4"
         style="background-color:#000000;border-bottom:1px solid #111827;">
 
@@ -54,16 +53,10 @@
                             Config
                         </a>
                     </li>
-                    {% if login_required %}
-                    <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
-                    </li>
-                    {% endif %}
                 </ul>
             </div>
         </div>
     </nav>
-    {% endif %}
 
     <main class="container mb-5">
         {% with messages = get_flashed_messages(with_categories=true) %}
