+++ templates/config.html
@@ -45,73 +45,6 @@
     <div class="col-12">
         <div class="card shadow-sm border-secondary">
             <div class="card-body">
-                <h2 class="card-title mb-3">Artillery settings</h2>
-
-                <p class="text-muted small mb-3">
-                    Configure Artillery application settings.
-                </p>
-
-                <form method="post" action="{{ url_for('config_page') }}">
-                    <div class="mb-3">
-                        <label class="form-label" for="log_lines_display">Log lines to display in output panel</label>
-                        <input type="number" id="log_lines_display" name="log_lines_display" 
-                               class="form-control" min="10" max="1000" 
-                               value="{{ artillery_config.log_lines_display }}" />
-                        <div class="form-text">
-                            Number of log lines to show in the real-time output panel (default: 50).
-                            Set higher to see more history, or lower for faster updates.
-                        </div>
-                    </div>
-
-                    <div class="mb-3">
-                        <label class="form-label" for="error_lines_display">Error lines to display</label>
-                        <input type="number" id="error_lines_display" name="error_lines_display" 
-                               class="form-control" min="5" max="200" 
-                               value="{{ artillery_config.error_lines_display }}" />
-                        <div class="form-text">
-                            Number of error lines to keep when viewing task errors (default: 20).
-                            Higher values show more error context.
-                        </div>
-                    </div>
-
-                    <div class="mb-3 form-check">
-                        <input type="checkbox" class="form-check-input" id="truncate_lines" 
-                               name="truncate_lines" value="on"
-                               {% if artillery_config.truncate_lines %}checked{% endif %} />
-                        <label class="form-check-label" for="truncate_lines">
-                            Enable line truncation
-                        </label>
-                        <div class="form-text">
-                            Truncate long lines in log output to the configured maximum length (default: enabled).
-                            Disable to see full lines without truncation.
-                        </div>
-                    </div>
-
-                    <div class="mb-3">
-                        <label class="form-label" for="max_line_length">Maximum line length</label>
-                        <input type="number" id="max_line_length" name="max_line_length" 
-                               class="form-control" min="50" max="1000" 
-                               value="{{ artillery_config.max_line_length }}" />
-                        <div class="form-text">
-                            Maximum number of characters per line (default: 200).
-                            Lines longer than this will be truncated with "..." when line truncation is enabled.
-                        </div>
-                    </div>
-
-                    <button type="submit" name="action" value="save_artillery" class="btn btn-danger">
-                        Save Artillery settings
-                    </button>
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- Media wall controls -->
-<div class="row mt-4">
-    <div class="col-12">
-        <div class="card shadow-sm border-secondary">
-            <div class="card-body">
                 <h2 class="card-title mb-3">Media wall</h2>
 
                 <p class="text-muted small mb-3">
