--- templates/login.html
@@ -1,41 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<!-- FSS-style animated background -->
-<div id="fss-container" class="fss-background">
-    <div id="fss-output" class="fss-output"></div>
-    <!-- optional controls element if needed: <div id="controls"></div> -->
-    <!-- background canvas will be injected here by JS -->
-    <noscript>
-        <style>.login-wrapper{position:relative;}</style>
-    </noscript>
-</div>
-
-<div class="login-wrapper d-flex align-items-center justify-content-center">
-    <div class="card login-card shadow-sm p-4">
-        <div class="text-center mb-3">
-            <div class="login-logo mx-auto mb-2">
-                <img src="{{ url_for('static', filename='logo.png') }}" alt="Artillery logo">
-            </div>
-            <h1 class="h4 mb-1 text-uppercase">Artillery</h1>
-            <p class="text-muted small mb-0">Sign in to continue</p>
-        </div>
-        <form method="post" class="login-form">
-            <div class="mb-3">
-                <label class="form-label" for="username">Username</label>
-                <input type="text" id="username" name="username" class="form-control"
-                    placeholder="Enter username" autocomplete="username" required>
-            </div>
-            <div class="mb-3">
-                <label class="form-label" for="password">Password</label>
-                <input type="password" id="password" name="password" class="form-control"
-                    placeholder="Enter password" autocomplete="current-password" required>
-            </div>
-            <button type="submit" class="btn btn-primary w-100 login-btn">
-                Login
-            </button>
-        </form>
-    </div>
-</div>
-<script src="{{ url_for('static', filename='js/login_fss.js') }}"></script>
-{% endblock %}
